Convert C++ to use double

+++ raytrace/cpp/camera.h
@@ -11,7 +11,7 @@ public:
         vertical = vec3(0.0, 2.0, 0.0);
         origin = vec3(0.0, 0.0, 0.0);
     }
-    ray get_ray(float u, float v) { return ray(origin, lower_left_corner + u*horizontal + v*vertical); }
+    ray get_ray(double u, double v) { return ray(origin, lower_left_corner + u*horizontal + v*vertical); }
 
     vec3 origin;
     vec3 lower_left_corner;
