+++ raytrace/cpp/hitable.h
@@ -4,14 +4,14 @@
 #include "ray.h"
 
 struct hit_record {
-    float t;
+    double t;
     vec3 p;
     vec3 normal;
 };
 
 class hitable {
 public:
-    virtual bool hit(const ray& r, float t_min, float t_max, hit_record& rec) const = 0;
+    virtual bool hit(const ray& r, double t_min, double t_max, hit_record& rec) const = 0;
 };
 
 #endif
